Обрезка фото и тг бот

+++ TGBot.py
@@ -1,6 +1,7 @@
 #!/usr/bin/python
 # -*- coding: cp1251  -*-
-from PreProcessPhoto import PreProcessPhoto
+from ImageUpdate import ImageUpdate
+from PreProcessPhoto import Otladka, PreProcessPhoto, PreProcessPhotov2
 import telebot;
 from telebot import types
 import cv2;
@@ -9,11 +10,24 @@ import os
 
 bot = telebot.TeleBot('6817178987:AAGbWfjbW9_GDZxSmQO-oloJPrj6_yHQxzM');
 PreProcess = True
-
+PreProcesser = PreProcessPhotov2()
 @bot.message_handler(content_types=['text'])  
-def get_text_messages(message):
-    
-    if message.text == "Обрезка фото":
+def get_text_messages(message): # Работа с сообщениями
+    if "thresh" in message.text:
+        PreProcesser.SetValues(pix1=PreProcesser.pix1, pix0 = PreProcesser.pix0,thresh=int(message.text.split()[1]))
+        print("set value thresh in " + message.text.split()[1])
+        return
+    if "pix0" in message.text:
+        PreProcesser.SetValues(pix0=int(message.text.split()[1]), pix1=PreProcesser.pix1, thresh=PreProcesser.thresh)
+        print("set value pix0 in " + message.text.split()[1])
+        return
+    if "pix1" in message.text:
+        PreProcesser.SetValues(pix1=int(message.text.split()[1]), pix0 = PreProcesser.pix0, thresh=PreProcesser.thresh)
+        print("set value pix1 in " + message.text.split()[1])
+        return
+    if "Values" == message.text:
+        print(str(PreProcesser.pix0), str(PreProcesser.pix1), str(PreProcesser.thresh)) 
+    if message.text == "Поиск документа":
         print("text with: " + str(message.from_user.id))
         bot.send_message(message.from_user.id, "Отправьте фото для обрезки P.S. Оберзаем фото документов на фоне другого цвета..")
         img = open("example.jpg", 'rb')
@@ -27,7 +41,7 @@ def get_text_messages(message):
     else:
         print("start with: " + str(message.from_user.id))
         markup = types.ReplyKeyboardMarkup(resize_keyboard=True)
-        btn1 = types.KeyboardButton("Обрезка фото")
+        btn1 = types.KeyboardButton("Поиск документа")
         btn2 = types.KeyboardButton("Распознавание текста")
         markup.add(btn1, btn2)
         bot.send_message(message.chat.id, text="Выбери опцию", reply_markup=markup)
@@ -39,20 +53,44 @@ def photo(message):
         ph.start()
     else:
          bot.send_message(message.from_user.id, "еще не робит..")
-def Work(message):
+def Work(message): # Работа с фото.
+    print("photo from: " + str(message.from_user.id))
+    bot.send_message(message.chat.id, text="Обработка...")
     fileID = message.photo[-1].file_id   
     file_info = bot.get_file(fileID)
     downloaded_file = bot.download_file(file_info.file_path)
     with open("temp" + str(message.from_user.id) +".jpg", 'wb') as new_file:
         new_file.write(downloaded_file)
-    photos = PreProcessPhoto.GetBestPhotos("temp" + str(message.from_user.id) +".jpg")
-    for item in photos:
-        cv2.imwrite("temp" + str(message.from_user.id) +".jpg",item)
+    photos = PreProcesser.Work("temp" + str(message.from_user.id) +".jpg")
+    if photos is None:
+        bot.send_message(message.chat.id, text="Документы не найдены")
+        return
+    # photos = ImageUpdate.Update(photos)
+    if len(photos) == 1:
+        cv2.imwrite("temp" + str(message.from_user.id) +".jpg",photos[0])
         img = open("temp" + str(message.from_user.id) +".jpg", 'rb')
-        bot.send_photo(message.from_user.id, img)
+        try:
+            bot.send_photo(message.from_user.id, img)
+        except:
+            bot.send_message(message.chat.id, text="Ошибка отправки")
+            print('ошибка отправки с  ' + str(message.from_user.id))
         img.close()
         print('Отправлено ' + str(message.from_user.id))
         os.remove("temp" + str(message.from_user.id) +".jpg")
+        return
+    else:
+        for item in photos:
+            cv2.imwrite("temp" + str(message.from_user.id) +".jpg",item)
+            img = open("temp" + str(message.from_user.id) +".jpg", 'rb')
+            try:
+                bot.send_photo(message.from_user.id, img)
+            except:
+                bot.send_message(message.chat.id, text="Ошибка отправки")
+                print('ошибка отправки с  ' + str(message.from_user.id))
+            img.close()
+            print('Отправлено ' + str(message.from_user.id))
+            os.remove("temp" + str(message.from_user.id) +".jpg")
+    return
 print("Start...")
 bot.polling(none_stop=True, interval=0)
